Consolidate identity verification declarations and apply code review advice
