+++ src/schema/v2/schema.ts
@@ -130,7 +130,7 @@ import { channel } from "./article/channel"
 import { departments, job, jobs } from "./jobs"
 import { RecentlySoldArtworks } from "./recentlySoldArtworks"
 import { artworksForUser } from "./artworksForUser"
-import { identityVerifications } from "./identityVerifications"
+import { identityVerificationsConnection } from "./identityVerification"
 
 const PrincipalFieldDirective = new GraphQLDirective({
   name: "principalField",
@@ -184,7 +184,7 @@ const rootFields = {
   geneFamiliesConnection: GeneFamilies,
   highlights: HighlightsField,
   homePage: HomePage,
-  identityVerifications,
+  identityVerificationsConnection,
   job,
   jobs,
   doNotUseImageSearch: ImageSearchField,
